StaffManagementSystem_Portable/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.4.0
pandas==2.0.3
numpy==1.24.3
gdown==4.7.1
requests==2.31.0
openpyxl==3.1.2
werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
rembg==2.0.50
onnxruntime==1.15.1
scipy==1.10.1
qrcode==7.4.2
opencv-python==4.8.1.78
cloudinary==1.36.0
gunicorn==21.2.0
requests==2.31.0
cryptography==41.0.7